✨ Add new features and improve existing functionality
- 🌐 **i18n**: Updated translations in `messages/en.json` and `messages/es.json`.
- 📦 **Dependencies**: Updated `package.json` and `pnpm-lock.yaml` to reflect dependency changes.
- 🗂️ **Database**:
	- Updated `prisma/schema.prisma` to modify the database schema.
	- Updated `prisma/ERD.svg` to reflect the new entity relationship diagram.
- 🖥️ **Frontend**:
	- Improved account and league pages (`src/app/[locale]/account/page.tsx`, `src/app/[locale]/league/page.tsx`).
	- Added a new reusable dialog component (`src/components/ui/dialog.tsx`).
	- Enhanced invite and league views (`src/components/ViewInvites.tsx`, `src/components/ViewLeague.tsx`).
- 🔒 **API**:
	- Added a new endpoint to kick users from a league (`src/app/api/league/[liga]/user/[user]/kick/route.ts`).
	- Improved invite handling (`src/app/api/league/[liga]/invite/accept/route.ts`, `src/app/api/league/[liga]/invite/create/route.ts`).
	- Enhanced league creation and deletion endpoints (`src/app/api/league/create/route.ts`, `src/app/api/league/delete/route.ts`).
	- Updated user-related API functionality (`src/app/api/user/route.ts`).
- 🔑 **Auth**: Refactored `src/auth.ts` for better security and maintainability.

- 🚀 These changes improve functionality, add new features, and enhance user experience.

+++ src/components/ViewInvites.tsx
@@ -6,14 +6,40 @@ import { redirect } from "next/navigation";
 import { toast } from "sonner"
 import { Skeleton } from "@/components/ui/skeleton";
 import Image from "next/image";
-// import { AlertDialog, AlertDialogAction, AlertDialogCancel, AlertDialogContent, AlertDialogDescription, AlertDialogFooter, AlertDialogHeader, AlertDialogTitle, AlertDialogTrigger, } from "@/components/ui/alert-dialog"
+import { AlertDialog, AlertDialogAction, AlertDialogCancel, AlertDialogContent, AlertDialogDescription, AlertDialogFooter, AlertDialogHeader, AlertDialogTitle, AlertDialogTrigger, } from "@/components/ui/alert-dialog"
+import { Dialog, DialogContent, DialogDescription, DialogHeader, DialogTitle, DialogTrigger, DialogFooter } from "@/components/ui/dialog"
+import { Input } from "@/components/ui/input"
+import { Label } from "@/components/ui/label"
+import { Button } from "@/components/ui/button"
 
 import type { Invitacion, Liga, User } from "@prisma/client"
 import { ApiKey } from "@/lib/utils";
 
-// async function createInvite(liga: Liga, id_user: string, usos: number) {
+async function createInvite(liga: Liga, id_user: string, usos: number) {
+	const res = await fetch(`/api/league/${liga.id}/invite/create`, {
+		method: 'POST',
+		headers: {
+			'Content-Type': 'application/json'
+		},
+		body: JSON.stringify({
+			id: id_user,
+			apiKey: ApiKey,
+			privada: liga.privada,
+			usos: usos
+		})
+	})
 
-// }
+	if (res.status === 200) {
+		return (await res.json()).invitacion as Invitacion
+	} else if (res.status === 400) {
+		const data = await res.json()
+		if (data.error === 'No changes made') {
+			return undefined
+		}
+	} else {
+		return false
+	}
+}
 
 async function deleteInvite(invite: Invitacion) {
 	const res = await fetch(`/api/league/${invite.id_liga}/invite/delete`, {
@@ -49,6 +75,8 @@ async function acceptInvite(invite: Invitacion, user: User) {
 
 	if (res.status === 200) {
 		return true
+	} else if (res.status === 403) {
+		return undefined
 	} else {
 		return false
 	}
@@ -92,6 +120,7 @@ export default function ViewInvites({ id }: { id: string }) {
 	const [invites, setInvites] = useState<Invitacion[]>([])
 	const [pendingUsers, setPendingUsers] = useState<{ id_invite: string, users: User[] }[]>()
 	const [loading, setLoading] = useState(true)
+	const [usos, setUsos] = useState<number>(0)
 	const [error, setError] = useState<string | null>(null)
 	const t = useTranslations('league.invites')
 	const locale = useLocale()
@@ -186,7 +215,66 @@ export default function ViewInvites({ id }: { id: string }) {
 			<div className="flex flex-col items-center justify-center bg-gray-300/40 dark:bg-gray-800/40 shadow-lg shadow-blue-400/40 dark:shadow-blue-900/50 mt-20 rounded-lg mx-2 md:mx-10">
 				<div className="flex items-center justify-center w-full gap-4 md:gap-6">
 					<h1 className="text-2xl font-semibold dark:text-white mt-2">{t('1', { name: liga.nombre })}</h1>
-					<button className="flex items-center justify-center mt-2 py-1 rounded-sm bg-green-600 text-white font-semibold px-2 hover:bg-green-700 transition duration-200 ease-in-out hover:cursor-pointer active:scale-[.97] transform">{t('13')}</button>
+					<Dialog>
+						<DialogTrigger asChild className="hover:cursor-pointer">
+							<Button className="flex items-center justify-center mt-2 py-1 rounded-sm bg-green-600 text-white font-semibold px-2 hover:bg-green-700 transition duration-200 ease-in-out hover:cursor-pointer active:scale-[.97] transform">{t('13')}</Button>
+						</DialogTrigger>
+						<DialogContent className="sm:max-w-[425px] dark:bg-gray-800 dark:text-white dark:border-gray-700">
+							<DialogHeader>
+								<DialogTitle>{t('13')}</DialogTitle>
+								<DialogDescription className="dark:text-white">
+									{t('25')}
+								</DialogDescription>
+							</DialogHeader>
+							<div className="grid gap-4 py-4">
+								<div className="grid grid-cols-4 items-center gap-4">
+									<Label htmlFor="name" className="text-right">
+										{t('26')}
+									</Label>
+									<Input id="name" className="col-span-3"
+										onChange={(e) => setUsos(parseInt(e.target.value))}
+										onWheel={(e) => {
+											if (parseInt(e.currentTarget.value) + (e.deltaY < 0 ? 1 : -1) > (liga.id_creador === session.user.id ? 999 : 10)) {
+												e.preventDefault()
+											} else if (parseInt(e.currentTarget.value) + (e.deltaY < 0 ? 1 : -1) < 1) {
+												e.preventDefault()
+											} else {
+												setUsos(parseInt(e.currentTarget.value) + (e.deltaY < 0 ? 1 : -1))
+											}
+										}}
+										value={usos || 1}
+										placeholder="1"
+										type="number"
+										min={1}
+										max={liga.id_creador === session.user.id ? 999 : 10}
+										step={1}
+										required={true}
+										autoFocus={true} />
+								</div>
+							</div>
+							<DialogFooter>
+								<Button
+									type="submit"
+									onClick={async (e) => {
+										if (usos > 0) {
+											const res = await createInvite(liga, session.user.id, usos)
+											if (res) {
+												toast.success(t('27'))
+												setInvites([...invites, { ...res }])
+											} else if (res === undefined) {
+												toast.error(t('30'))
+											} else {
+												toast.error(t('28'))
+											}
+										} else {
+											toast.error(t('29'))
+										}
+									}}
+									className="hover:cursor-pointer active:scale-[.97] transforom duration-200 dark:bg-blue-800 dark:hover:bg-blue-900"
+								>{t('13')}</Button>
+							</DialogFooter>
+						</DialogContent>
+					</Dialog>
 				</div>
 				<div className="flex items-center justify-center w-full">
 					{
@@ -198,7 +286,7 @@ export default function ViewInvites({ id }: { id: string }) {
 										{i > 0 && <hr className="w-3/4 border-gray-400 dark:border-gray-600 my-4" />}
 										<div className="flex flex-col items-center justify-center w-full">
 											<div className="flex items-center justify-center w-full gap-4 md:gap-6 mt-5">
-												<div className="flex items-center justify-center dark:text-white md:w-3/6 mb-2 gap-2">
+												<div className="flex items-center justify-center dark:text-white md:w-5/6 mb-2 gap-2">
 													<span className="font-semibold md:w-3/6">
 														{i + 1}.{' '}
 														<input
@@ -213,20 +301,34 @@ export default function ViewInvites({ id }: { id: string }) {
 															navigator.clipboard.writeText(`${process.env.NEXT_PUBLIC_INVITE_URL}/${invite.id}`)
 															toast.success(t('2'))
 														}}
-														className="flex items-center justify-center w-1/5 py-2 rounded-sm bg-blue-500 text-white font-semibold px-2 hover:bg-blue-600 transition duration-200 ease-in-out hover:cursor-pointer active:scale-[.97] transform"
+														className="flex items-center justify-center py-2 rounded-sm bg-blue-500 text-white font-semibold px-2 hover:bg-blue-600 transition duration-200 ease-in-out hover:cursor-pointer active:scale-[.97] transform"
 													>{t('11')}</button>
-													<button
-														onClick={async () => {
-															const res = await deleteInvite(invite)
-															if (res) {
-																toast.success(t('18'))
-																setInvites(invites.filter((i) => i.id !== invite.id))
-															} else {
-																toast.error(t('19'))
-															}
-														}}
-														className="flex items-center w-1/5 justify-center py-2 rounded-sm bg-red-500 text-white font-semibold px-2 hover:bg-red-600 transition duration-200 ease-in-out hover:cursor-pointer active:scale-[.97] transform"
-													>{t('14')}</button>
+													<AlertDialog>
+														<AlertDialogTrigger className="bg-red-500 text-white px-4 py-2 rounded-md hover:bg-red-600 hover:cursor-pointer active:scale-[.97] transform duration-200 font-semibold">{t('14')}</AlertDialogTrigger>
+														<AlertDialogContent className="dark:bg-gray-900 bg-gray-300 dark:text-white">
+															<AlertDialogHeader>
+																<AlertDialogTitle>{t('15')}</AlertDialogTitle>
+																<AlertDialogDescription className="dark:text-white">{t('16')}</AlertDialogDescription>
+															</AlertDialogHeader>
+															<AlertDialogFooter>
+																<AlertDialogCancel
+																	className="hover:cursor-pointer active:scale-[.97] transform duration-200"
+																>{t('17')}</AlertDialogCancel>
+																<AlertDialogAction
+																	className="bg-red-500 text-white px-4 py-2 rounded-md hover:bg-red-600 hover:cursor-pointer active:scale-[.97] transform duration-200"
+																	onClick={async () => {
+																		const res = await deleteInvite(invite)
+																		if (res) {
+																			toast.success(t('18'))
+																			setInvites(invites.filter((i) => i.id !== invite.id))
+																		} else {
+																			toast.error(t('19'))
+																		}
+																	}}
+																>{t('14')}</AlertDialogAction>
+															</AlertDialogFooter>
+														</AlertDialogContent>
+													</AlertDialog>
 													<p className="dark:text-white w-1/5 text-center rounded-sm hidden md:inline">{t('12', { usos: invite.usosRestantes })}</p>
 												</div>
 											</div>
@@ -259,7 +361,9 @@ export default function ViewInvites({ id }: { id: string }) {
 																								const res = await acceptInvite(invite, u)
 																								if (res) {
 																									toast.success(t('23'))
-																									setInvites(invites.filter((i) => i.id !== invite.id))
+																									redirect(`/${locale}/league/${liga.id}/invites`)
+																								} else if (res === undefined) {
+																									toast.error(t('31'))
 																								} else {
 																									toast.error(t('24'))
 																								}
@@ -296,7 +400,7 @@ export default function ViewInvites({ id }: { id: string }) {
 
 							})
 						) : (
-							<p>No hay</p>
+							<h2 className="dark:text-white my-4 font-semibold">{t('32')}</h2>
 						)
 					}
 				</div>
